added new enum type for proabablity attribute; refactoring;
