+++ toshi_hazard_store/model/attributes/__init__.py
@@ -5,4 +5,6 @@ from .attributes import (
     LevelValuePairAttribute,
     PickleAttribute,
 )
-from .unicode_enum_constrained_attribute import UnicodeEnumConstrainedAttribute
+from .unicode_enum_constrained_attribute import EnumAttribute, UnicodeEnumConstrainedAttribute
+
+# from .enum_unicode_attribute import EnumUnicodeAttribute
